pyproject.toml
[tool.poetry]
name = "airflow-yandex-api"
version = "0.1.0"
description = ""
authors = ["Leonide"]
readme = "README.md"
packages = [{include = "airflow_yandex_api"}]

[tool.poetry.dependencies]
python = "^3.10"
pandas = "1.5.3"
python-dotenv = "0.21.1"
requests = "2.28.2"
apache-airflow = "2.5.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
